requirements.txt
flask
jinja2
flask-login
flask-session
peewee
authlib
requests
python-dotenv
